Add warning flags to Makefile

+++ project/src/common/constants.hpp
@@ -13,7 +13,7 @@
 #define PACKET_ID_LEN (3)
 
 #define PLAYER_ID_MAX_LEN (6)
-#define PLAYER_ID_MAX (pow(10, PLAYER_ID_MAX_LEN) - 1)
+#define PLAYER_ID_MAX ((int)pow(10, PLAYER_ID_MAX_LEN) - 1)
 
 #define FILE_BUFFER_LEN (512)
